+++ src/algorithms/tracking/libs/dll_pll_conf.h
@@ -41,6 +41,8 @@ public:
     double fs_in{2000000.0};
     double carrier_lock_th{0.0};
     double bs_dominance_ratio{0.6};
+    double bs_runner_up_margin{0.2};
+    double bs_transition_confidence{0.6};
     float pll_pull_in_bw_hz{50.0};
     float dll_pull_in_bw_hz{3.0};
     float fll_bw_hz{35.0};
@@ -74,6 +76,8 @@ public:
     int32_t max_carrier_lock_fail{0};
     int32_t bs_stable_best_required{3};
     int32_t bs_min_events_for_lock{10};
+    int32_t bs_transition_window_epochs{3};
+    int32_t bs_tentative_events_required{2};
     char signal[3]{};
     char system{'G'};
     bool enable_fll_pull_in{false};