+++ src/algorithms/tracking/libs/dll_pll_conf_fpga.h
@@ -44,6 +44,8 @@ public:
     double fs_in{12500000.0};
     double carrier_lock_th{0.0};
     double bs_dominance_ratio{0.6};
+    double bs_runner_up_margin{0.2};
+    double bs_transition_confidence{0.6};
 
     float pll_pull_in_bw_hz{50.0};
     float dll_pull_in_bw_hz{3.0};
@@ -85,6 +87,8 @@ public:
     // int32_t max_lock_fail;
     int32_t bs_stable_best_required{3};
     int32_t bs_min_events_for_lock{10};
+    int32_t bs_transition_window_epochs{3};
+    int32_t bs_tentative_events_required{2};
 
     int32_t* ca_codes{nullptr};
     int32_t* data_codes{nullptr};